refactor(ai-builder): Fix high priority code review issues
- Fix potential crash from null assertion on lsClient
- Remove debug: true from LangSmith client in production
- Extract duplicate PairwiseTargetOutput type to shared types/pairwise.ts
- Improve isPairwiseTargetOutput type guard to validate all properties
