+++ packages/@n8n/ai-workflow-builder.ee/evaluations/types/pairwise.ts
@@ -1,0 +1,39 @@
+import type { EvaluationResult as LangsmithEvaluationResult } from 'langsmith/evaluation';
+
+import type { EvalCriteria } from '../utils/judge-panel';
+
+// ============================================================================
+// Dataset Input/Output Types
+// ============================================================================
+
+export interface PairwiseDatasetInput {
+	evals: {
+		dos: string;
+		donts: string;
+	};
+	prompt: string;
+	/** Optional notion_id from dataset for tracing */
+	notion_id?: string;
+}
+
+export interface PairwiseTargetOutput {
+	prompt: string;
+	evals: EvalCriteria;
+	/** Pre-computed feedback results. Named with underscore to avoid LangSmith auto-processing */
+	_feedback: LangsmithEvaluationResult[];
+}
+
+// ============================================================================
+// Type Guards
+// ============================================================================
+
+export function isPairwiseTargetOutput(outputs: unknown): outputs is PairwiseTargetOutput {
+	if (!outputs || typeof outputs !== 'object') return false;
+	const obj = outputs as Record<string, unknown>;
+	return (
+		typeof obj.prompt === 'string' &&
+		Array.isArray(obj._feedback) &&
+		obj.evals !== undefined &&
+		typeof obj.evals === 'object'
+	);
+}
